refactor: extract common combo box functionality
